+++ src/octokit/octokit.ts
@@ -1,26 +1,39 @@
-import { Octokit } from "@octokit/rest";
-import { loadUser, loadRepos, logout, loadUserAndRepos } from "../store/helpers";
-import { storage } from '../storage';
-import { configs } from '../configs';
+import { Octokit } from '@octokit/rest';
+import { notLoggedUser, reloadRepos } from '../store/helpers';
 import vscode from 'vscode';
+import { Auth } from './auth';
+import { dataStore } from '../store';
 
 export let octokit: Octokit | null = null;
 export let token = '';
 
+let mayBeLogged = false;
 /**
  * Automatically enters the token, if .env or stored token.
  *
  * @export
  */
-export function activateOctokit(): void {
-  let token = '';
-  if (process.env.USE_ENV_TOKEN === 'true')
-    token = process.env.TOKEN;
-  else if (configs.saveToken)
-    token = storage.loadToken();
+export async function activateOctokit(): Promise<void> {
+  try {
+    let token;
 
-  if (token)
-    initOctokit(token);
+    if (process.env.USE_ENV_TOKEN === 'true')
+      token = process.env.TOKEN;
+
+    token = await Auth.init();
+
+    // Init octokit if we have a local stored token
+    if (token)
+      initOctokit(token);
+
+    else
+      dataStore.dispatch({ type: 'UPDATE_USER', value: notLoggedUser() });
+
+  } catch (err) {
+    vscode.window.showErrorMessage(err.message);
+    console.error('activateOctokit error: ', err);
+    dataStore.dispatch({ type: 'UPDATE_USER', value: notLoggedUser() });
+  }
 }
 
 export async function initOctokit(tokenArg: string): Promise<void> {
@@ -28,22 +41,23 @@ export async function initOctokit(tokenArg: string): Promise<void> {
     auth: tokenArg,
   });
 
+  // As the user may logout just after logging in, we add this to prevent the error box.
+  mayBeLogged = true;
   try {
-    await loadUserAndRepos();
-  }
-
-  catch (err) {
-    vscode.window.showErrorMessage(err.message);
-    console.error('Octokit init error: ', err);
+    await reloadRepos();
+  } catch (err) {
+    if (mayBeLogged) {
+      vscode.window.showErrorMessage(err.message);
+      console.error('Octokit init error: ', err);
+    }
     octokit = null;
     return;
   }
   token = tokenArg;
-  if (configs.saveToken) // 'If setting', store the token.
-    storage.storeToken(tokenArg);
 }
 
 export function logoutAndForgetToken(): void {
-  storage.removeToken();
-  logout();
+  mayBeLogged = false;
+  dataStore.dispatch({ type: 'UPDATE_USER', value: notLoggedUser() });
+  Auth.logout();
 }