+++ src/storage.ts
@@ -1,15 +1,30 @@
 import vscode from 'vscode';
-import { configs } from './configs';
-
-export namespace storage {
-
-  let context: vscode.ExtensionContext;
-
-  const keys = {
-    token: 'token'
-  };
 
 
+let context: vscode.ExtensionContext;
+
+// https://stackoverflow.com/a/57857305
+function remove(key: string) {
+  context.globalState.update(key, undefined);
+}
+
+// TODO support for default value
+// not using getter/setter/proxy as they may be async.
+class Item<T>{
+  constructor(private key: string) { }
+  get(): T | undefined {
+    // for some reason the undefined type isn't being automatically added.
+    return context.globalState.get<T>(this.key);
+  }
+  set(value: T) {
+    return context.globalState.update(this.key, value);
+  }
+  remove() {
+    remove(this.key);
+  }
+}
+
+class StorageClass {
   /**
    * Passes the context, and also removes the current stored token if saveToken configuration is false.
    * For this token behavior, this must be executed before octokit stuff.
@@ -17,35 +32,22 @@ export namespace storage {
    * @export
    * @param {vscode.ExtensionContext} contextArg
    */
-  export function activate(contextArg: vscode.ExtensionContext) {
+  activate(contextArg: vscode.ExtensionContext) {
     context = contextArg;
 
-    if (!configs.saveToken && loadToken())
-      removeToken();
-  }
-
-  // https://stackoverflow.com/a/57857305
-  function removeKey(key: string) {
-    context.globalState.update(key, undefined);
+    // Legacy. Leave it here for some time to remove stored tokens.
+    if (this.token.get())
+      this.token.remove();
   }
 
   // Removes all keys
-  export function resetGlobalState() {
+  resetGlobalState() {
     // Forces the read of the private _value.
     const keys = Object.keys((context.globalState as any)._value);
-    keys.forEach(key => removeKey(key));
+    keys.forEach(key => remove(key));
   }
 
-  export function storeToken(token: string) {
-    return context.globalState.update(keys.token, token);
-  }
-
-  export function loadToken() {
-    return context.globalState.get<string>(keys.token);
-  }
-
-  // Or could export removeKey and keys.
-  export function removeToken() {
-    removeKey(keys.token);
-  }
+  token = new Item<string>('token');
 }
+
+export const Storage = new StorageClass();