+++ src/treeView/repositories/repositories.ts
@@ -1,7 +1,9 @@
+// Tree view base
+
 import vscode from 'vscode';
 import { BaseTreeDataProvider } from '../base';
-import DataStore from '../../store';
-import { reloadRepos } from "../../store/helpers";
+import { dataStore } from '../../store';
+import { reloadRepos } from '../../store/helpers';
 import { RepoItem } from './repoItem';
 import { getClonedTreeItem, activateClonedRepos } from './clonedRepos';
 import { activateNotClonedRepos, getNotClonedTreeItem } from './notClonedRepos';
@@ -13,14 +15,14 @@ export let repositoriesTreeDataProvider: TreeDataProvider;
 export function activateTreeViewRepositories() {
   repositoriesTreeDataProvider = new TreeDataProvider();
 
-  DataStore.subscribe(() => { repositoriesTreeDataProvider.refresh(); });
+  dataStore.subscribe(() => { repositoriesTreeDataProvider.refresh(); });
 
   vscode.window.registerTreeDataProvider('githubRepoMgr.views.repositories',
     repositoriesTreeDataProvider);
 
   // Access GitHub Web Page
   vscode.commands.registerCommand('githubRepoMgr.commands.repos.openWebPage', ({ repo }: RepoItem) =>
-    vscode.commands.executeCommand("vscode.open", vscode.Uri.parse(repo.url)));
+    vscode.commands.executeCommand('vscode.open', vscode.Uri.parse(repo.url)));
 
   // Will have .git on the end.
   vscode.commands.registerCommand('githubRepoMgr.commands.repos.copyRepositoryUrl', ({ repo }: RepoItem) =>