+++ src/uiCommands/uiCloneTo.ts
@@ -1,7 +1,7 @@
 import { workspace, window, Uri, commands } from 'vscode';
 import { Repository } from '../store/types';
 import { reloadRepos } from '../store/helpers';
-import { cloneRepo } from "../octokit/commands/cloneRepo";
+import { cloneRepo } from '../octokit/commands/cloneRepo';
 import path from 'path';
 import { configs } from '../configs';
 
@@ -68,12 +68,13 @@ export async function uiCloneTo(repo: Repository) {
 
   const action = await window.showInformationMessage(`Cloned ${repo.name} to ${repoPath}!`,
     openStr, openInNewWindowStr, addToWorkspaceStr);
+
   switch (action) {
-    case openStr:
-      commands.executeCommand('vscode.openFolder', uri); break;
-    case openInNewWindowStr:
-      commands.executeCommand('vscode.openFolder', uri, true); break;
-    case addToWorkspaceStr:
-      workspace.updateWorkspaceFolders(workspace.workspaceFolders!.length, 0, { uri }); break;
+  case openStr:
+    commands.executeCommand('vscode.openFolder', uri); break;
+  case openInNewWindowStr:
+    commands.executeCommand('vscode.openFolder', uri, true); break;
+  case addToWorkspaceStr:
+    workspace.updateWorkspaceFolders(workspace.workspaceFolders.length, 0, { uri }); break;
   }
-};
+}